Split admin/project/ actions into separate actions files for each ones
